totp settings #3
added in the logic for enabling totp for your account. also added in
some placeholder ui for updating your password.

+++ db/postgres/init.sql
@@ -57,7 +57,7 @@ create table authn_totp (
     users_id bigint primary key not null references users (id),
     algo int not null,
     step int not null,
-    digits int not null,
+    digits smallint not null,
     secret bytea not null
 );
